requirements.txt
flask
flask-cors
requests
python-dotenv
pymongo
bcrypt
dnspython
